pyproject.toml
[project]
name = "gradescope-mcp"
version = "0.1.0"
description = "MCP server for Gradescope course management"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.0",
    "gradescopeapi>=1.7.0",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.2",
]

[project.scripts]
gradescope-mcp = "gradescope_mcp.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gradescope_mcp"]

[dependency-groups]
dev = [
    "pytest>=9.0",
]
